Respond the webhook only based on payload parsing. Actual Event proccessing is move to backround thread.
This require some refactoring, created ReciveWebhook and ReciveEventFile
functions to represent the different behavior in Web Server VS CLI
triggering while keeping to the GH stuff in the GH package
